refactor: restructure path_query_push into standard v0/v1 versioned dispatch
Move the implementation into element/path_query_push/{mod.rs,v0.rs,v1.rs}:
v0 freezes the legacy limit/offset accounting for GROVE_V1..GROVE_V3, v1
(GROVE_V4+) carries the issue #690 fix as its own implementation, and the
ElementQueryExtensions::path_query_push trait method becomes the match
dispatcher, replacing the inline version check. Also document the gate in
the GROVE_V4 module doc.
