+++ grovedb/src/element/query.rs
@@ -1,8 +1,8 @@
 //! Query
 //! Implements functions in Element for querying
 use grovedb_costs::{
-    cost_return_on_error, cost_return_on_error_into, cost_return_on_error_into_no_add,
-    cost_return_on_error_no_add, CostResult, CostsExt, OperationCost,
+    cost_return_on_error, cost_return_on_error_into_no_add, cost_return_on_error_no_add,
+    CostResult, CostsExt, OperationCost,
 };
 use grovedb_element::Element;
 use grovedb_merk::{
@@ -12,10 +12,7 @@ use grovedb_merk::{
 };
 use grovedb_path::SubtreePath;
 use grovedb_storage::{rocksdb_storage::RocksDbStorage, RawIterator, StorageContext};
-use grovedb_version::{
-    check_grovedb_v0, check_grovedb_v0_or_v1_with_cost, check_grovedb_v0_with_cost,
-    version::GroveVersion,
-};
+use grovedb_version::{check_grovedb_v0, check_grovedb_v0_with_cost, version::GroveVersion};
 
 use crate::{
     element::{path_query_push_args::PathQueryPushArgs, query_options::QueryOptions},
@@ -328,229 +325,27 @@ impl ElementQueryExtensions for Element {
     }
 
     /// Push arguments to path query
+    ///
+    /// Version dispatch — see the `path_query_push` module: v0 is the legacy
+    /// limit/offset accounting frozen for `GROVE_V1`..`GROVE_V3`; v1
+    /// (`GROVE_V4`+) no longer charges the outer limit for subqueries
+    /// emptied by offset skips (issue #690).
     fn path_query_push(
         args: PathQueryPushArgs,
         grove_version: &GroveVersion,
     ) -> CostResult<(), Error> {
-        use crate::util::{compat, TxRef};
-
-        let path_query_push_version = grove_version.grovedb_versions.element.path_query_push;
-        check_grovedb_v0_or_v1_with_cost!("path_query_push", path_query_push_version);
-
-        // println!("path_query_push {} \n", args);
-
-        let mut cost = OperationCost::default();
-
-        let PathQueryPushArgs {
-            storage,
-            transaction,
-            key,
-            element,
-            path,
-            subquery_path,
-            subquery,
-            left_to_right,
-            query_options,
-            result_type,
-            results,
-            limit,
-            offset,
-        } = args;
-
-        let tx = TxRef::new(storage, transaction);
-
-        let QueryOptions {
-            allow_get_raw,
-            allow_cache,
-            decrease_limit_on_range_with_no_sub_elements,
-            ..
-        } = query_options;
-        if element.is_any_tree() {
-            let mut path_vec = path.to_vec();
-            let key = cost_return_on_error_no_add!(
-                cost,
-                key.ok_or(Error::MissingParameter(
-                    "the key must be provided when using a subquery path",
-                ))
-            );
-            path_vec.push(key);
-
-            if let Some(subquery) = subquery {
-                if let Some(subquery_path) = &subquery_path {
-                    path_vec.extend(subquery_path.iter().map(|k| k.as_slice()));
-                }
-
-                let inner_query = SizedQuery::new(subquery, *limit, *offset);
-                let path_vec_owned = path_vec.iter().map(|x| x.to_vec()).collect();
-                let inner_path_query = PathQuery::new(path_vec_owned, inner_query);
-
-                let (mut sub_elements, skipped) = cost_return_on_error!(
-                    &mut cost,
-                    Element::get_path_query(
-                        storage,
-                        &inner_path_query,
-                        query_options,
-                        result_type,
-                        transaction,
-                        grove_version,
-                    )
-                );
-
-                if let Some(limit) = limit {
-                    if sub_elements.is_empty()
-                        && decrease_limit_on_range_with_no_sub_elements
-                        && (path_query_push_version == 0 || skipped == 0)
-                    {
-                        // we should decrease by 1 in this case
-                        *limit = limit.saturating_sub(1);
-                    } else {
-                        *limit =
-                            limit.saturating_sub(sub_elements.len().min(u16::MAX as usize) as u16);
-                    }
-                }
-                if let Some(offset) = offset {
-                    *offset = offset.saturating_sub(skipped);
-                }
-                results.append(&mut sub_elements.elements);
-            } else if let Some(subquery_path) = subquery_path {
-                if offset.unwrap_or(0) == 0 {
-                    if let Some((subquery_path_last_key, subquery_path_front_keys)) =
-                        &subquery_path.split_last()
-                    {
-                        path_vec.extend(subquery_path_front_keys.iter().map(|k| k.as_slice()));
-
-                        let subtree_path: SubtreePath<_> = path_vec.as_slice().into();
-                        let subtree = cost_return_on_error!(
-                            &mut cost,
-                            compat::merk_optional_tx(
-                                storage,
-                                subtree_path,
-                                tx.as_ref(),
-                                None,
-                                grove_version
-                            )
-                        );
-
-                        match result_type {
-                            QueryElementResultType => {
-                                if let Some(element) = cost_return_on_error_into!(
-                                    &mut cost,
-                                    Element::get_optional_with_absolute_refs(
-                                        &subtree,
-                                        path_vec.as_slice(),
-                                        subquery_path_last_key.as_slice(),
-                                        allow_cache,
-                                        grove_version,
-                                    )
-                                ) {
-                                    results.push(QueryResultElement::ElementResultItem(element));
-                                }
-                            }
-                            QueryKeyElementPairResultType => {
-                                if let Some(element) = cost_return_on_error_into!(
-                                    &mut cost,
-                                    Element::get_optional_with_absolute_refs(
-                                        &subtree,
-                                        path_vec.as_slice(),
-                                        subquery_path_last_key.as_slice(),
-                                        allow_cache,
-                                        grove_version,
-                                    )
-                                ) {
-                                    results.push(QueryResultElement::KeyElementPairResultItem((
-                                        subquery_path_last_key.to_vec(),
-                                        element,
-                                    )));
-                                }
-                            }
-                            QueryPathKeyElementTrioResultType => {
-                                if let Some(element) = cost_return_on_error_into!(
-                                    &mut cost,
-                                    Element::get_optional_with_absolute_refs(
-                                        &subtree,
-                                        path_vec.as_slice(),
-                                        subquery_path_last_key.as_slice(),
-                                        allow_cache,
-                                        grove_version,
-                                    )
-                                ) {
-                                    results.push(QueryResultElement::PathKeyElementTrioResultItem(
-                                        (
-                                            path_vec.iter().map(|p| p.to_vec()).collect(),
-                                            subquery_path_last_key.to_vec(),
-                                            element,
-                                        ),
-                                    ));
-                                }
-                            }
-                        }
-                    } else {
-                        return Err(Error::CorruptedCodeExecution(
-                            "subquery_paths can not be empty",
-                        ))
-                        .wrap_with_cost(cost);
-                    };
-
-                    if let Some(limit) = limit {
-                        *limit = limit.saturating_sub(1);
-                    }
-                } else if let Some(offset) = offset {
-                    *offset = offset.saturating_sub(1);
-                }
-            } else if allow_get_raw {
-                cost_return_on_error_no_add!(
-                    cost,
-                    Element::basic_push(
-                        PathQueryPushArgs {
-                            storage,
-                            transaction,
-                            key: Some(key),
-                            element,
-                            path,
-                            subquery_path,
-                            subquery,
-                            left_to_right,
-                            query_options,
-                            result_type,
-                            results,
-                            limit,
-                            offset,
-                        },
-                        grove_version
-                    )
-                );
-            } else {
-                return Err(Error::InvalidPath(
-                    "you must provide a subquery or a subquery_path when interacting with a Tree \
-                     of trees"
-                        .to_owned(),
-                ))
-                .wrap_with_cost(cost);
-            }
-        } else {
-            cost_return_on_error_no_add!(
-                cost,
-                Element::basic_push(
-                    PathQueryPushArgs {
-                        storage,
-                        transaction,
-                        key,
-                        element,
-                        path,
-                        subquery_path,
-                        subquery,
-                        left_to_right,
-                        query_options,
-                        result_type,
-                        results,
-                        limit,
-                        offset,
-                    },
-                    grove_version
-                )
-            );
+        match grove_version.grovedb_versions.element.path_query_push {
+            0 => crate::element::path_query_push::path_query_push_v0(args, grove_version),
+            1 => crate::element::path_query_push::path_query_push_v1(args, grove_version),
+            version => Err(Error::VersionError(
+                grovedb_version::error::GroveVersionError::UnknownVersionMismatch {
+                    method: "path_query_push".to_string(),
+                    known_versions: vec![0, 1],
+                    received: version,
+                },
+            ))
+            .wrap_with_cost(OperationCost::default()),
         }
-        Ok(()).wrap_with_cost(cost)
     }
 
     /// Takes a sized query and a key and returns subquery key and subquery as
